requirements.txt
numpy
torchvision
#https://download.pytorch.org/whl/cpu/torch-1.0.1.post2-cp37-cp37m-linux_x86_64.whl
fastai
starlette
uvicorn==0.3.32
python-multipart
aiofiles
aiohttp
google-cloud-storage
